Model, ModelManager: refactored to use drink rather than person or addressbook
